requirements.txt
dashscope==1.23.1
elastic-transport==8.17.1
elasticsearch==8.18.1
langchain==0.3.7
langchain-chroma==0.2.4
langchain-community==0.3.7
langchain-core==0.3.63
langchain-elasticsearch==0.3.2
langchain-experimental==0.3.4
langchain-mcp-adapters==0.1.7
langchain-openai==0.2.3
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.0
langgraph-sdk==0.1.70
langsmith==0.1.147

openai==1.66.3
opencv-python==4.11.0.86
pandas==2.2.3
python-docx==1.1.2
python-dotenv==1.1.0

Flask==2.3.3
Flask-CORS==4.0.0
minio==7.1.15

# AWS SDK for MinIO 兼容支持
boto3==1.34.3
botocore==1.34.3

SQLAlchemy==2.0.23
pymysql==1.1.0

numpy==1.26.4
python-multipart==0.0.20
python-docx==1.1.2
PyPDF2==3.0.1
regex==2024.11.6

fastapi~=0.109.1
requests~=2.32.0
uvicorn~=0.26.0
pydantic~=1.10.12